+++ CHANGELOG.md
@@ -1,2 +1,11 @@
-## 0.0.1
+## 24.0.0-dev.1.2
 
+Added getLicenseExpiryDate(), refactored example app, added new example in main.dart.
+
+## 24.0.0-dev.1.1
+
+Added dartdoc and updated package description.
+
+## 24.0.0-dev.1
+
+This is a prerelease for Anyline 24 only supporting Android.